#36 periodType in container, remove unnecessary periodic/actualMeterReads
